Added the crioctl tool
This is just a temporary solution to be able to test checkpoint/restore
functionality in CRI-O as long as the checkpoint/restore CRI API changes
are not merged in Kubernetes.

The crioctl tool currently provides two commands:

 * crioctl checkpoint
 * crioctl restore

These two commands are using the same parameters as the
checkpoint/restore functionality in Podman.

With this temporary tool it is possible to test the checkpoint/restore
implementation in CRI-O before it is merged in Kubernetes and it also
enables users to checkpoint containers from Kubernetes and restore the
containers somewhere else.

Signed-off-by: Adrian Reber <areber@redhat.com>